Beginning to merge add user function
