+++ .gitignore
@@ -5,3 +5,4 @@
 /dist
 scratch.js
 bundleStats.html
+src/components/Form/HomeScreenMOCK.jsx
