+++ src/components/Form/ManageUsers.jsx
@@ -3,6 +3,11 @@ import { useSession } from '@inrupt/solid-ui-react';
 import Container from '@mui/material/Container';
 import Paper from '@mui/material/Paper';
 import Box from '@mui/material/Box';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogContentText from '@mui/material/DialogContentText';
+import DialogTitle from '@mui/material/DialogTitle';
 import { useStatusNotification, useField } from '../../hooks';
 import { runNotification, addUserToPod } from '../../utils';
 import FormSection from './FormSection';
@@ -22,6 +27,13 @@ const ManageUsers = () => {
   const { clearValue: clearUserName, ...userName } = useField('text');
   const { clearValue: clearUserUrl, ...userUrl } = useField('text');
   const { setUserList } = useContext(UserListContext);
+
+  const [open, setOpen] = React.useState(false);
+  const [open2, setOpen2] = React.useState(false);
+  // const handleOpen = () => setOpen(true);
+  // const handleOpen2 = () => setOpen2(true);
+  const handleClose = () => setOpen(false);
+  const handleClose2 = () => setOpen2(false);
 
   // Event handler for adding user from users list
   const handleAddUser = async (event) => {
@@ -65,46 +77,58 @@ const ManageUsers = () => {
   };
 
   return (
-    <Container component="" maxWidth="">
-      <Box
-        sx={{
-          margin: 1,
-          display: 'flex',
-          flexDirection: 'column',
-          padding: '5px',
-          width: '95vw'
-        }}
-      >
-        <Paper elevation={2} sx={{ display: 'inline-block', mx: '2px', padding: '20px' }}>
-          <FormSection
-            title="Manage Users"
-            state={state}
-            statusType="Status"
-            defaultMessage="To be added..."
+    <Dialog open={open2} onClose={handleClose2}>
+      <DialogTitle align="center">Add new client</DialogTitle>
+      <DialogContent>
+        <Container component="" maxWidth="">
+          <DialogContentText>Enter new client details below</DialogContentText>
+          <Box
+            sx={{
+              margin: 1,
+              display: 'flex',
+              flexDirection: 'column',
+              padding: '5px',
+              width: '95vw'
+            }}
           >
-            <form onSubmit={handleAddUser} style={formRowStyle} autoComplete="off">
-              <div>
-                <label htmlFor="add-user-name">User's name (i.e. John Doe): </label>
-                <input id="add-user-name" name="addUserName" {...userName} />{' '}
-              </div>
-              <br />
-              <div>
-                <label htmlFor="add-user-url">
-                  Add Pod URL to users list (i.e., username.opencommons.net):{' '}
-                </label>
-                <br />
-                <br />
-                <input id="add-user-url" name="addUserUrl" size="60" {...userUrl} />{' '}
-              </div>
-              <br />
-              <button type="submit" disabled={state.processing}>
-                Add User
-              </button>
-            </form>
-          </FormSection>
-        </Paper>
-      </Box>
-    </Container>
+            <Paper elevation={2} sx={{ display: 'inline-block', mx: '2px', padding: '20px' }}>
+              <FormSection
+                title="Manage Users"
+                state={state}
+                statusType="Status"
+                defaultMessage="To be added..."
+              >
+                <form onSubmit={handleAddUser} style={formRowStyle} autoComplete="off">
+                  <div>
+                    <label htmlFor="add-user-name">User's name (i.e. John Doe): </label>
+                    <input id="add-user-name" name="addUserName" {...userName} />{' '}
+                  </div>
+                  <br />
+                  <div>
+                    <label htmlFor="add-user-url">
+                      Add Pod URL to users list (i.e., username.opencommons.net):{' '}
+                    </label>
+                    <br />
+                    <br />
+                    <input id="add-user-url" name="addUserUrl" size="60" {...userUrl} />{' '}
+                  </div>
+                  <br />
+                  <button type="submit" disabled={state.processing}>
+                    Add User
+                  </button>
+                </form>
+              </FormSection>
+            </Paper>
+          </Box>
+        </Container>
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={handleClose2}>Cancel</Button>
+        <Button onClick={handleSubmit} variant="contained">
+          Submit
+        </Button>
+      </DialogActions>
+    </Dialog>
   );
 };
